--- src/main/java/EtfRecommendService/notification/NotificationDto.java
@@ -1,10 +1,0 @@
-package EtfRecommendService.notification;
-
-public record NotificationDto(
-        String message,
-        ReceiverType receiverType,
-        NotificationType type,
-        String targetId
-) {
-
-}
