+++ src/main/java/EtfRecommendService/notification/dto/NotificationDto.java
@@ -1,0 +1,13 @@
+package EtfRecommendService.notification.dto;
+
+import EtfRecommendService.notification.NotificationType;
+import EtfRecommendService.notification.ReceiverType;
+
+public record NotificationDto(
+        String message,
+        ReceiverType receiverType,
+        NotificationType type,
+        String targetId
+) {
+
+}
